Ensure clean shutdown and in flight messages are completed
Whilst running the compoenent with DEBUG turned on I observed the
component never clears the setInterval timer.

A 2nd behaviour observed and experienced in proeuction is that when
stop() is called the azure client is immediately stopped and closed,
this results in any messages currently being processed by a handler to
fail even if the actual processing within the handler succeeded, this is
becuase the call with the handler `await receiver.completeMessage(brokeredMessage);`
can never complete as the connection is closed. This results in the messagew
being requeued and retried.

This chnage tracks open subscriptions using a new array openSubscriptions then when
stop is called each openSubscription is closed first. Based on MS docs this first
blocks and drains the open subscription cleanly. The code is then able to cleanly
wait on any inflight messages still be processed before closing the azure client.

Tests all pass (although sometimes, they are a bit flaky).