feat: add subject field to comply the change in the azure lib (#71)
It seems that adding a `label` property is no longer supported by the Azure Service Bus library. We have found (not in the documentation) that now the lib is expecting a property called `subject`. 
This change adds a `subject` property next to the existing `label` one (just in case of any potential backwards issue).